requirements.txt
google-genai>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
websockets>=12.0
playwright>=1.40.0
requests>=2.31.0
aiohttp>=3.9.0
